I_Piece, I_Boardに分けて作成

+++ Shogi3DBattle/header/B_NaturalBufferedData.h
@@ -27,6 +27,6 @@ public:
         return S_OK;
     }
 
-    unsigned int GetSize () override           {return _datas.size();};
+    unsigned int GetSize () override           {return static_cast<unsigned int>(_datas.size());};
     void         SetDatas(std::vector<T> datas){_datas = datas;}       // データ集合セット
 };
